+++ eventmesh-connector-plugin/eventmesh-connector-pravega/src/main/java/org/apache/eventmesh/connector/pravega/SubscribeTask.java
@@ -1,16 +1,15 @@
-package org.apache.eventmesh.connector.pravega.client;
-
-import org.apache.eventmesh.api.EventListener;
-import org.apache.eventmesh.api.EventMeshAction;
-import org.apache.eventmesh.api.EventMeshAsyncConsumeContext;
-
-import java.util.concurrent.atomic.AtomicBoolean;
+package org.apache.eventmesh.connector.pravega;
 
 import io.cloudevents.CloudEvent;
 import io.pravega.client.stream.EventRead;
 import io.pravega.client.stream.EventStreamReader;
-
 import lombok.extern.slf4j.Slf4j;
+import org.apache.eventmesh.api.EventListener;
+import org.apache.eventmesh.api.EventMeshAction;
+import org.apache.eventmesh.api.EventMeshAsyncConsumeContext;
+import org.apache.eventmesh.connector.pravega.client.PravegaEvent;
+
+import java.util.concurrent.atomic.AtomicBoolean;
 
 @Slf4j
 public class SubscribeTask extends Thread {
